@@ -7,64 +7,61 @@ import static io.gatling.javaapi.core.CoreDsl.jsonPath;
 import static io.gatling.javaapi.http.HttpDsl.http;
 
 public class ClerkHttpService extends HttpService {
+
 	public HttpRequestActionBuilder generateOMRSIdentifier() {
 		return http("Generate OMRS Identifier")
-				.post("/openmrs/ws/rest/v1/idgen/identifiersource/8549f706-7e85-4c1d-9424-217d50a2988b/identifier")
-				.body(StringBody("{}"))
-				.check(jsonPath("$.identifier").saveAs("identifier"));
+		        .post("/openmrs/ws/rest/v1/idgen/identifiersource/8549f706-7e85-4c1d-9424-217d50a2988b/identifier")
+		        .body(StringBody("{}")).check(jsonPath("$.identifier").saveAs("identifier"));
 	}
-	
+
 	public HttpRequestActionBuilder sendPatientRegistrationRequest() {
 		String registrationRequestTemplate = """
-					{
-					   "identifiers":[
-					      {
-					         "identifier":"#{identifier}",
-					         "identifierType":"05a29f94-c0ed-11e2-94be-8c13b969e334",
-					         "location":"44c3efb0-2583-4c80-a79e-1f756a03c0a1",
-					         "preferred":true
-					      }
-					   ],
-					   "person":{
-					      "gender":"M",
-					      "age":47,
-					      "birthdate":"1970-01-01T00:00:00.000+0100",
-					      "birthdateEstimated":false,
-					      "dead":false,
-					      "deathDate":null,
-					      "causeOfDeath":null,
-					      "names":[
-					         {
-					            "givenName":"Jayasanka",
-					            "familyName":"Smith"
-					         }
-					      ],
-					      "addresses": [
-					        {
-					        "address1": "30, Vivekananda Layout, Munnekolal,Marathahalli",
-					        "cityVillage": "Bengaluru",
-					        "country": "India",
-					        "postalCode": "560037"
-					        }
-					      ]
-					    }
-					}
-					
-			""";
-		
-		return http("Send Patient Registration Request")
-				.post("/openmrs/ws/rest/v1/patient/")
-				.body(StringBody(registrationRequestTemplate))
-				.check(jsonPath("$.uuid").saveAs("patientUuid"));
+		        		{
+		        		   "identifiers":[
+		        		      {
+		        		         "identifier":"#{identifier}",
+		        		         "identifierType":"05a29f94-c0ed-11e2-94be-8c13b969e334",
+		        		         "location":"44c3efb0-2583-4c80-a79e-1f756a03c0a1",
+		        		         "preferred":true
+		        		      }
+		        		   ],
+		        		   "person":{
+		        		      "gender":"M",
+		        		      "age":47,
+		        		      "birthdate":"1970-01-01T00:00:00.000+0100",
+		        		      "birthdateEstimated":false,
+		        		      "dead":false,
+		        		      "deathDate":null,
+		        		      "causeOfDeath":null,
+		        		      "names":[
+		        		         {
+		        		            "givenName":"Jayasanka",
+		        		            "familyName":"Smith"
+		        		         }
+		        		      ],
+		        		      "addresses": [
+		        		        {
+		        		        "address1": "30, Vivekananda Layout, Munnekolal,Marathahalli",
+		        		        "cityVillage": "Bengaluru",
+		        		        "country": "India",
+		        		        "postalCode": "560037"
+		        		        }
+		        		      ]
+		        		    }
+		        		}
+
+		        """;
+
+		return http("Send Patient Registration Request").post("/openmrs/ws/rest/v1/patient/")
+		        .body(StringBody(registrationRequestTemplate)).check(jsonPath("$.uuid").saveAs("patientUuid"));
 	}
-	
+
 	public HttpRequestActionBuilder getPersonAttributeType(String personAttributeTypeUuid) {
-		return http("Get Person Attribute Type")
-				.get("/openmrs/ws/rest/v1/personattributetype/"+personAttributeTypeUuid);
+		return http("Get Person Attribute Type").get("/openmrs/ws/rest/v1/personattributetype/" + personAttributeTypeUuid);
 	}
-	
+
 	public HttpRequestActionBuilder getOrderedAddressHierarchyLevels() {
 		return http("Get Ordered Address Hierarchy Levels")
-				.get("/openmrs/module/addresshierarchy/ajax/getOrderedAddressHierarchyLevels.form");
+		        .get("/openmrs/module/addresshierarchy/ajax/getOrderedAddressHierarchyLevels.form");
 	}
 }
