@@ -6,16 +6,14 @@ import org.openmrs.performance.scenarios.VisitPatientScenario;
 
 import java.util.List;
 
-public class DoctorPersona extends Persona<DoctorRegistry>{
-	
+public class DoctorPersona extends Persona<DoctorRegistry> {
+
 	public DoctorPersona(double loadShare) {
 		super(loadShare);
 	}
-	
+
 	@Override
 	public List<Scenario<DoctorRegistry>> getScenarios() {
-		return List.of(
-			new VisitPatientScenario(1)
-		);
+		return List.of(new VisitPatientScenario(1));
 	}
 }
