@@ -6,11 +6,12 @@ import org.openmrs.performance.scenarios.Scenario;
 import java.util.List;
 
 public abstract class Persona<R extends Registry<?>> {
+
 	public double loadShare;
-	
+
 	public Persona(double loadShare) {
 		this.loadShare = loadShare;
 	}
-	
+
 	public abstract List<Scenario<R>> getScenarios();
 }
